changed doubles to int

+++ src/rotation_magic.c
@@ -6,34 +6,34 @@
 /*   By: phtruong <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/08/29 18:21:50 by phtruong          #+#    #+#             */
-/*   Updated: 2019/08/30 13:05:51 by phtruong         ###   ########.fr       */
+/*   Updated: 2019/09/05 16:09:02 by phtruong         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../includes/ft_fdf.h"
 
-void	fdf_rot_x(double *y, int *z, double alpha)
+void	fdf_rot_x(int *y, int *z, double alpha)
 {
-	double prev_y;
+	int prev_y;
 
 	prev_y = *y;
 	*y = prev_y * cos(alpha) + *z * sin(alpha);
 	*z = -prev_y * sin(alpha) + *z * cos(alpha);
 }
 
-void	fdf_rot_y(double *x, int *z, double beta)
+void	fdf_rot_y(int *x, int *z, double beta)
 {
-	double prev_x;
+	int prev_x;
 
 	prev_x = *x;
 	*x = prev_x * cos(beta) + *z * sin(beta);
 	*z = -prev_x * sin(beta) + *z * cos(beta);
 }
 
-void	fdf_rot_z(double *x, double *y, double eta)
+void	fdf_rot_z(int *x, int *y, double eta)
 {
-	double prev_x;
-	double prev_y;
+	int prev_x;
+	int prev_y;
 
 	prev_x = *x;
 	prev_y = *y;
@@ -41,10 +41,10 @@ void	fdf_rot_z(double *x, double *y, double eta)
 	*y = prev_x * sin(eta) + prev_y * cos(eta);
 }
 
-void	fdf_iso(double *x, double *y, int z)
+void	fdf_iso(int *x, int *y, int z)
 {
-	double prev_x;
-	double prev_y;
+	int prev_x;
+	int prev_y;
 
 	prev_x = *x;
 	prev_y = *y;
